trimester/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.3.0
Pillow==10.1.0
matplotlib==3.8.2
